#218: Refactor ProductReferenceResolverTest into separate reference resolver test classes.
