+++ src/test/java/com/commercetools/sync/products/helpers/productreferenceresolver/TaxCategoryReferenceResolverTest.java
@@ -1,0 +1,196 @@
+package com.commercetools.sync.products.helpers.productreferenceresolver;
+
+import com.commercetools.sync.commons.exceptions.ReferenceResolutionException;
+import com.commercetools.sync.products.ProductSyncOptions;
+import com.commercetools.sync.products.ProductSyncOptionsBuilder;
+import com.commercetools.sync.products.helpers.ProductReferenceResolver;
+import com.commercetools.sync.services.TaxCategoryService;
+import io.sphere.sdk.client.SphereClient;
+import io.sphere.sdk.models.LocalizedString;
+import io.sphere.sdk.models.Reference;
+import io.sphere.sdk.models.SphereException;
+import io.sphere.sdk.products.ProductDraftBuilder;
+import io.sphere.sdk.products.ProductVariantDraft;
+import io.sphere.sdk.producttypes.ProductType;
+import io.sphere.sdk.taxcategories.TaxCategory;
+import org.junit.Before;
+import org.junit.Test;
+
+import javax.annotation.Nonnull;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.concurrent.CompletableFuture;
+
+import static com.commercetools.sync.commons.MockUtils.getMockCategoryService;
+import static com.commercetools.sync.commons.MockUtils.getMockTypeService;
+import static com.commercetools.sync.inventories.InventorySyncMockUtils.getMockChannelService;
+import static com.commercetools.sync.inventories.InventorySyncMockUtils.getMockSupplyChannel;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getMockProductService;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getMockProductTypeService;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getMockStateService;
+import static com.commercetools.sync.products.ProductSyncMockUtils.getMockTaxCategoryService;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.when;
+
+public class TaxCategoryReferenceResolverTest {
+    private static final String CHANNEL_KEY = "channel-key_1";
+    private static final String CHANNEL_ID = "1";
+    private static final String PRODUCT_TYPE_ID = "productTypeId";
+    private static final String TAX_CATEGORY_ID = "taxCategoryId";
+    private static final String STATE_ID = "stateId";
+    private static final String PRODUCT_ID = "productId";
+
+    private ProductReferenceResolver referenceResolver;
+    private TaxCategoryService taxCategoryService;
+
+    /**
+     * Sets up the services and the options needed for reference resolution.
+     */
+    @Before
+    public void setup() {
+        taxCategoryService = getMockTaxCategoryService(TAX_CATEGORY_ID);
+        final ProductSyncOptions syncOptions = ProductSyncOptionsBuilder.of(mock(SphereClient.class)).build();
+        referenceResolver = new ProductReferenceResolver(syncOptions,
+            getMockProductTypeService(PRODUCT_TYPE_ID), getMockCategoryService(), getMockTypeService(),
+            getMockChannelService(getMockSupplyChannel(CHANNEL_ID, CHANNEL_KEY)), taxCategoryService,
+            getMockStateService(STATE_ID), getMockProductService(PRODUCT_ID));
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithKeysAsUuidSetAndAllowed_ShouldResolveReference() {
+        final ProductSyncOptions productSyncOptions = ProductSyncOptionsBuilder.of(mock(SphereClient.class))
+                                                                               .allowUuidKeys(true)
+                                                                               .build();
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(TaxCategory.referenceOfId(UUID.randomUUID().toString()));
+
+        final ProductReferenceResolver productReferenceResolver = new ProductReferenceResolver(productSyncOptions,
+            getMockProductTypeService(PRODUCT_TYPE_ID), getMockCategoryService(), getMockTypeService(),
+            getMockChannelService(getMockSupplyChannel(CHANNEL_ID, CHANNEL_KEY)), taxCategoryService,
+            getMockStateService(STATE_ID), getMockProductService(PRODUCT_ID));
+
+        final ProductDraftBuilder resolvedDraft = productReferenceResolver.resolveTaxCategoryReferences(productBuilder)
+                                                                          .toCompletableFuture().join();
+
+        assertThat(resolvedDraft.getTaxCategory()).isNotNull();
+        assertThat(resolvedDraft.getTaxCategory().getId()).isEqualTo(TAX_CATEGORY_ID);
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithKeys_ShouldResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(TaxCategory.referenceOfId("taxCategoryKey"));
+
+        final ProductDraftBuilder resolvedDraft = referenceResolver.resolveTaxCategoryReferences(productBuilder)
+                                                                   .toCompletableFuture().join();
+
+        assertThat(resolvedDraft.getTaxCategory()).isNotNull();
+        assertThat(resolvedDraft.getTaxCategory().getId()).isEqualTo(TAX_CATEGORY_ID);
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithKeysAsUuidSetAndNotAllowed_ShouldNotResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(TaxCategory.referenceOfId(UUID.randomUUID().toString()))
+            .key("dummyKey");
+
+        assertThat(referenceResolver.resolveTaxCategoryReferences(productBuilder).toCompletableFuture())
+            .hasFailed()
+            .hasFailedWithThrowableThat()
+            .isExactlyInstanceOf(ReferenceResolutionException.class)
+            .hasMessage("Failed to resolve reference 'tax-category' on ProductDraft"
+                + " with key:'" + productBuilder.getKey() + "'. Reason: Found a UUID"
+                + " in the id field. Expecting a key without a UUID value. If you want to"
+                + " allow UUID values for reference keys, please use the "
+                + "allowUuidKeys(true) option in the sync options.");
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithNullTaxCategory_ShouldNotResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .key("dummyKey");
+
+        assertThat(referenceResolver.resolveTaxCategoryReferences(productBuilder).toCompletableFuture())
+            .hasNotFailed()
+            .isCompletedWithValueMatching(resolvedDraft -> Objects.isNull(resolvedDraft.getTaxCategory()));
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithNonExistentTaxCategory_ShouldNotResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(TaxCategory.referenceOfId("nonExistentKey"))
+            .key("dummyKey");
+
+        when(taxCategoryService.fetchCachedTaxCategoryId(anyString()))
+            .thenReturn(CompletableFuture.completedFuture(Optional.empty()));
+
+        assertThat(referenceResolver.resolveTaxCategoryReferences(productBuilder).toCompletableFuture())
+            .hasNotFailed()
+            .isCompletedWithValueMatching(resolvedDraft ->
+                Objects.nonNull(resolvedDraft.getTaxCategory())
+                    && Objects.equals(resolvedDraft.getTaxCategory().getId(), "nonExistentKey"));
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithNullIdOnTaxCategoryReference_ShouldNotResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(Reference.of(TaxCategory.referenceTypeId(), (String)null))
+            .key("dummyKey");
+
+        assertThat(referenceResolver.resolveTaxCategoryReferences(productBuilder).toCompletableFuture())
+            .hasFailed()
+            .hasFailedWithThrowableThat()
+            .isExactlyInstanceOf(ReferenceResolutionException.class)
+            .hasMessage("Failed to resolve reference 'tax-category' on ProductDraft with "
+                + "key:'" + productBuilder.getKey() + "'. Reason: Reference 'id' field"
+                + " value is blank (null/empty).");
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithEmptyIdOnTaxCategoryReference_ShouldNotResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(TaxCategory.referenceOfId(""))
+            .key("dummyKey");
+
+        assertThat(referenceResolver.resolveTaxCategoryReferences(productBuilder).toCompletableFuture())
+            .hasFailed()
+            .hasFailedWithThrowableThat()
+            .isExactlyInstanceOf(ReferenceResolutionException.class)
+            .hasMessage("Failed to resolve reference 'tax-category' on ProductDraft with "
+                + "key:'" + productBuilder.getKey() + "'. Reason: Reference 'id' field"
+                + " value is blank (null/empty).");
+    }
+
+    @Test
+    public void resolveTaxCategoryReference_WithExceptionOnTaxCategoryFetch_ShouldNotResolveReference() {
+        final ProductDraftBuilder productBuilder = getBuilderWithRandomProductTypeUuid()
+            .taxCategory(TaxCategory.referenceOfId("taxCategoryKey"))
+            .key("dummyKey");
+
+        final CompletableFuture<Optional<String>> futureThrowingSphereException = new CompletableFuture<>();
+        futureThrowingSphereException.completeExceptionally(new SphereException("CTP error on fetch"));
+        when(taxCategoryService.fetchCachedTaxCategoryId(anyString())).thenReturn(futureThrowingSphereException);
+
+        assertThat(referenceResolver.resolveTaxCategoryReferences(productBuilder).toCompletableFuture())
+            .hasFailed()
+            .hasFailedWithThrowableThat()
+            .isExactlyInstanceOf(SphereException.class)
+            .hasMessageContaining("CTP error on fetch");
+    }
+
+    @Nonnull
+    private static ProductDraftBuilder getBuilderWithProductTypeRefId(@Nonnull final String refId) {
+        return ProductDraftBuilder.of(ProductType.referenceOfId(refId),
+            LocalizedString.ofEnglish("testName"),
+            LocalizedString.ofEnglish("testSlug"),
+            (ProductVariantDraft)null);
+    }
+
+    @Nonnull
+    private static ProductDraftBuilder getBuilderWithRandomProductTypeUuid() {
+        return getBuilderWithProductTypeRefId(UUID.randomUUID().toString());
+    }
+}
